pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rtgs-lab-tools"
version = "0.1.0"
description = "Environmental sensing data tools, gridded climate data access, and IoT device management for RTGS Lab"
authors = [
    {name = "RTGS Lab", email = "rtgs@umn.edu"},
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
]
dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "sqlalchemy>=1.4.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=0.19.0",
    "requests>=2.25.0",
    "click>=8.0.0",
    "pyarrow>=10.0.0",
    "matplotlib>=3.5.0",
]

[project.optional-dependencies]
climate = [
    "xarray>=2022.3.0",
    "cdsapi>=0.5.0",
    "netcdf4>=1.6.0",
]
visualization = [
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
]
mcp = [
    "mcp[cli]>=0.1.0",
    "fastmcp>=0.1.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=0.991",
    "pre-commit>=2.20.0",
]
all = [
    "rtgs-lab-tools[climate,visualization,mcp,dev]",
]

[project.scripts]
rtgs = "rtgs_lab_tools.cli:cli"

[project.urls]
Homepage = "https://github.com/RTGS-Lab/gems_sensing_db_tools"
Repository = "https://github.com/RTGS-Lab/gems_sensing_db_tools"
Documentation = "https://rtgs.umn.edu/"

[tool.hatch.build.targets.wheel]
packages = ["src/rtgs_lab_tools"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  venv
  | \.venv
  | \.git
  | \.mypy_cache
  | \.pytest_cache
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
src_paths = ["src", "tests"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_reexport = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "psycopg2.*",
    "cdsapi.*",
    "mcp.*",
    "fastmcp.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=src/rtgs_lab_tools --cov-report=term-missing --cov-report=html"